rsinter: integrated get_forward_events into intersink
If the property is not set, no changes will be made to the
producer's settings (to avoid clobbering whatever the default
might be over time).  When prepare() is called, and if the
forward events is not set, then it is read from the producer
so that the element's property matches the configuration.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>